Lessons 42 no errros  May 8

+++ backend/src/resolvers/Mutation.js
@@ -1,19 +1,20 @@
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
-const { randomBytes } = require('crypto');
-const {promisify} = require('util');
-const { transport, makeANiceEmail } = require('../mail')
-
+const { randomBytes } = require("crypto");
+const { promisify } = require("util");
+const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    if(!ctx.request.userId){
-      throw new Error('You must be logged on to do that');
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to do that!");
     }
+
     const item = await ctx.db.mutation.createItem(
       {
         data: {
-          //this is how to create relationship between the Item and the users
+          // This is how to create a relationship between the Item and the User
           user: {
             connect: {
               id: ctx.request.userId
@@ -30,11 +31,11 @@ const Mutations = {
     return item;
   },
   updateItem(parent, args, ctx, info) {
-    //first take a copy of the updates
+    // first take a copy of the updates
     const updates = { ...args };
-    //remove the ID from the updates
+    // remove the ID from the updates
     delete updates.id;
-    //run the update method
+    // run the update method
     return ctx.db.mutation.updateItem(
       {
         data: updates,
@@ -47,10 +48,16 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    // 1. find the Item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    // 1. find the item
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own this Item, or have the permissions.
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermission) {
+      throw new Error("You don't have permission to do that!");
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -58,7 +65,7 @@ const Mutations = {
   //////////////////SignUP////////////////////////
 
   async signup(parent, args, ctx, info) {
-    // to lower case
+    // lower case
     args.email = args.email.toLowerCase();
     //hash their password
     const password = await bcrypt.hash(args.password, 10);
@@ -107,44 +114,45 @@ const Mutations = {
   },
   signout(parent, args, ctx, info) {
     ctx.response.clearCookie("token");
-     return { message: "Goodbye!" };
+    return { message: "Goodbye!" };
   },
   async requestReset(parent, args, ctx, info) {
     //1. Check if this is real user
-    const user = await ctx.db.query.user({ where:{ email: args.email  }});
-    if(!user) {
-      throw new Error(`No such user found with this ${args.email}`)
+    const user = await ctx.db.query.user({ where: { email: args.email } });
+    if (!user) {
+      throw new Error(`No such user found with this ${args.email}`);
     }
     //2. Set a reset token and expiry on that user
     const randomBytesPromiseified = promisify(randomBytes);
-    const resetToken = (await randomBytesPromiseified(20)).toString('hex');
+    const resetToken = (await randomBytesPromiseified(20)).toString("hex");
     const resetTokenExpiry = Date.now() + 3600000; // This is 1 hour
     const res = await ctx.db.mutation.updateUser({
-      where: { email: args.email},
+      where: { email: args.email },
       data: { resetToken, resetTokenExpiry }
     });
 
     //3. Email them a reset token
     const mailRes = await transport.sendMail({
-    from: 'stephmen@gmail.com',
-    to: user.email,
-    subject: 'Your Password Reset Token',
-    html: makeANiceEmail(`Your Password Token is Here!
+      from: "stephmen@gmail.com",
+      to: user.email,
+      subject: "Your Password Reset Token",
+      html: makeANiceEmail(`Your Password Token is Here!
     \n\n
-    <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Clik Here to Reset</a>`)
-    })
-
+    <a href='${
+      process.env.FRONTEND_URL
+    }/reset?resetToken=${resetToken}'>Clik Here to Reset</a>`)
+    });
 
     //4. Return the message
-        return {message: "Thanks!"};
+    return { message: "Thanks!" };
   },
 
-///////////////////Reset Password ///////////////////
+  ///////////////////Reset Password ///////////////////
 
-  async resetPassword(parent, args, ctx, info){
+  async resetPassword(parent, args, ctx, info) {
     //1. check if the password match
-    if(args.password !== args.confirmPassword) {
-      throw new Error("Your Passwords dont match!")
+    if (args.password !== args.confirmPassword) {
+      throw new Error("Your Passwords dont match!");
     }
     //2.check if this is a legit resset tokem
     //3. check if the token is expired
@@ -152,32 +160,126 @@ const Mutations = {
       where: {
         resetToken: args.resetToken,
         resetTokenExpiry_gte: Date.now() - 36000000
+      }
+    });
+    if (!user) {
+      throw new Error("This token is either invalid or expired");
     }
-  })
-  if(!user) {
-    throw new Error('This token is either invalid or expired')
-  }
     //4. hash their new password
     const password = await bcrypt.hash(args.password, 10);
     //5. save the new passoword to the user and remove old reset token
     const updatedUser = await ctx.db.mutation.updateUser({
-       where: { email: user.email },
-       data: {
-         password,
-         resetToken: null,
-         resetTokenExpiry: null,
-       },
-    })
+      where: { email: user.email },
+      data: {
+        password,
+        resetToken: null,
+        resetTokenExpiry: null
+      }
+    });
     //6 generte JWT
-    const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET )
+    const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
     //7. set the JWt cookie
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365
-    })
+    });
     //8. return the new user
-    return updatedUser
+    return updatedUser;
   },
+  async updatePermissions(parent, args, ctx, info) {
+    //1. Check if the current user is logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must not logged in!");
+    }
+    //2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    //3. Check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    //4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
+  },
+  async addToCart(parents, args, ctx, info) {
+    //1. Make sure user is signed in
+    const { userId } = ctx.request;
+    if (!userId) {
+      throw new Error("you must be singed in soon");
+    }
+    //2. Query user current car
+    const [existingCartItem] = await ctx.db.query.cartItems({
+      where: {
+        user: { id: userId },
+        item: { id: args.id }
+      }
+    });
+    //3. Check if this current item is in the cart and if yes increment by one
+    if (existingCartItem) {
+      console.log("This item is already in their cart");
+      return ctx.db.mutation.updateCartItem(
+        {
+          where: { id: existingCartItem.id },
+          data: { quantity: existingCartItem.quantity + 1 }
+        },
+        info
+      );
+    }
+    //4. If the item is not in the cart create a new fresh item in the cart.
+    return ctx.db.mutation.createCartItem(
+      {
+        data: {
+          user: {
+            connect: { id: userId }
+          },
+          item: {
+            connect: { id: args.id }
+          }
+        }
+      },
+      info
+    );
+  },
+  async removeFromCart(parent, args, ctx, info) {
+    //1. Find the cart Item
+    const cartItem = await ctx.db.query.cartItem(
+      {
+        where: {
+          id: args.id
+        }
+      },
+      `{id, user { id }}`
+    );
+    //1.5 Make sure we found an Item
+    if (!cartItem) throw new Error("No Cart Item Found!");
+    //2 Make sure they own that cart Item
+    if (cartItem.user.id !== ctx.request.userId) {
+      throw new Error("Cheatin huhhhhhh");
+    }
+    //3. Delete that cart Item
+    return ctx.db.mutation.deleteCartItem(
+      {
+        where: { id: args.id }
+      },
+      info
+    );
+  }
 };
 
 module.exports = Mutations;
